Firebase stub: new implementation in place
